% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Dataset}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance of Fire.}
  \item{DAMAGE}{Damage Produced By Fire.}
  ...
}
}
\source{
Retrieved from firedam.csv provided on the Lab 7 Canvas Tab
}
\usage{
fire
}
\description{
A brief description of the dataset and its context. Explain what the dataset represents,
how it was collected, and any relevant information that a user might need to understand
or use the data.
}
\keyword{datasets}
