% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{distribution}

\item{fun}{function i.e., mean, sd, var}

\item{alpha}{1 - interval estimate}

\item{cx}{scale of text}

\item{...}{additional arguments}
}
\value{
a bootstrap with accomanying plot
}
\description{
Title
}
\examples{
myboot2(iter=10000,x = rnorm(20,mean=10,sd=4),fun="mean",alpha=0.05,cx=1.5)
}
