% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Binomial}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{p}{probability}

\item{...}{elipsses}
}
\value{
A Histogram of a Binomial
}
\description{
Binomial
}
\examples{
mycltb(n=20,iter=10000,p=0.5)
}
