% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Title}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{population vector}

\item{mu}{mean of population vector}

\item{sig}{standard deviation of population vector}

\item{...}{additional arguments}
}
\value{
a plot of the maximum likelihood estimates for the population mean and standard deviation
}
\description{
Title
}
\examples{
mymlnorm(x=c(10,10,11,12,12,13,16,15,15),mu=seq(10,15,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
