% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{the probability that X is lesser or equal to a}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
the probability density that the value of a variable that follows the standard normal distribution is equal to x
}
\description{
My Normal Curve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
