% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{My Plot}
\usage{
myplot(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
heights for inputs "x" by performing on it quadratic operations
}
\description{
My Plot
}
\examples{
myplot(32)
}
