% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{My Sample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{sample size}

\item{iter}{the amount of times sample will be produced}

\item{time}{time interval between sample & barplot production}
}
\value{
iter number of samples and barplots
}
\description{
My Sample
}
\examples{
mysample(n = 10, iter = 15, time = 1.5)
}
